Refactor: Move RandomNumberBallsWithButton function from MainActivity to HomeScreen
