+++ app/src/main/java/com/bonustrack02/lotterygenerator/MainActivity.kt
@@ -131,9 +131,8 @@ fun LotteryBall(
     number: Int,
     modifier: Modifier = Modifier
 ) {
-    // 색상 결정 로직
     val ballColor = when (number) {
-        in 1..10 -> LotteryYellow // 기존에 정의된 색상 사용
+        in 1..10 -> LotteryYellow
         in 11..20 -> LotteryBlue
         in 21..30 -> LotteryRed
         in 31..40 -> LotteryGray
@@ -143,7 +142,7 @@ fun LotteryBall(
     Box(
         contentAlignment = Alignment.Center,
         modifier = modifier
-            .size(40.dp) // 공 크기
+            .size(40.dp)
             .background(color = ballColor, shape = CircleShape)
     ) {
         Text(
@@ -152,70 +151,5 @@ fun LotteryBall(
             fontWeight = FontWeight.Bold,
             style = MaterialTheme.typography.bodyMedium
         )
-    }
-}
-
-@Composable
-fun RandomNumberBallsWithButton(
-    numbers: List<Int>,
-    onGenerateClick: () -> Unit,
-    modifier: Modifier = Modifier
-) {
-    Column(
-        horizontalAlignment = Alignment.CenterHorizontally,
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(16.dp)
-    ) {
-        Row(
-            horizontalArrangement = Arrangement.spacedBy(8.dp),
-            modifier = Modifier
-                .padding(bottom = 16.dp)
-                .height(40.dp)
-        ) {
-            if (numbers.isNotEmpty()) {
-                numbers.forEach { number ->
-                    Box(
-                        contentAlignment = Alignment.Center,
-                        modifier = Modifier
-                            .size(40.dp)
-                            .background(
-                                color = when (number) {
-                                    in 1..10 -> LotteryYellow
-                                    in 11..20 -> LotteryBlue
-                                    in 21..30 -> LotteryRed
-                                    in 31..40 -> LotteryGray
-                                    else -> LotteryGreen
-                                },
-                                shape = CircleShape
-                            )
-                    ) {
-                        Text(
-                            text = number.toString(),
-                            color = Color.White,
-                            fontWeight = FontWeight.Bold
-                        )
-                    }
-                }
-            } else {
-                repeat(6) {
-                    Box(
-                        modifier = Modifier
-                            .size(40.dp)
-                            .background(
-                                color = Color.LightGray.copy(alpha = 0.3f),
-                                shape = CircleShape
-                            )
-                    )
-                }
-            }
-        }
-
-        Button(
-            onClick = onGenerateClick,
-            modifier = Modifier.padding(top = 16.dp)
-        ) {
-            Text(stringResource(R.string.generate_new_number_set))
-        }
     }
 }
