+++ app/src/main/java/com/bonustrack02/lotterygenerator/presentation/home/HomeScreen.kt
@@ -1,16 +1,33 @@
 package com.bonustrack02.lotterygenerator.presentation.home
 
+import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import com.bonustrack02.lotterygenerator.RandomNumberBallsWithButton
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import com.bonustrack02.lotterygenerator.R
+import com.bonustrack02.lotterygenerator.ui.theme.LotteryBlue
+import com.bonustrack02.lotterygenerator.ui.theme.LotteryGray
+import com.bonustrack02.lotterygenerator.ui.theme.LotteryGreen
+import com.bonustrack02.lotterygenerator.ui.theme.LotteryRed
+import com.bonustrack02.lotterygenerator.ui.theme.LotteryYellow
 
 @Composable
 fun LotteryBallScreen(viewModel: HomeViewModel, modifier: Modifier = Modifier) {
@@ -31,6 +48,71 @@ fun LotteryBallScreen(viewModel: HomeViewModel, modifier: Modifier = Modifier) {
                 numbers = numbers,
                 onGenerateClick = { viewModel.generateNewNumbers() }
             )
+        }
+    }
+}
+
+@Composable
+fun RandomNumberBallsWithButton(
+    numbers: List<Int>,
+    onGenerateClick: () -> Unit,
+    modifier: Modifier = Modifier
+) {
+    Column(
+        horizontalAlignment = Alignment.CenterHorizontally,
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(16.dp)
+    ) {
+        Row(
+            horizontalArrangement = Arrangement.spacedBy(8.dp),
+            modifier = Modifier
+                .padding(bottom = 16.dp)
+                .height(40.dp)
+        ) {
+            if (numbers.isNotEmpty()) {
+                numbers.forEach { number ->
+                    Box(
+                        contentAlignment = Alignment.Center,
+                        modifier = Modifier
+                            .size(40.dp)
+                            .background(
+                                color = when (number) {
+                                    in 1..10 -> LotteryYellow
+                                    in 11..20 -> LotteryBlue
+                                    in 21..30 -> LotteryRed
+                                    in 31..40 -> LotteryGray
+                                    else -> LotteryGreen
+                                },
+                                shape = CircleShape
+                            )
+                    ) {
+                        Text(
+                            text = number.toString(),
+                            color = Color.White,
+                            fontWeight = FontWeight.Bold
+                        )
+                    }
+                }
+            } else {
+                repeat(6) {
+                    Box(
+                        modifier = Modifier
+                            .size(40.dp)
+                            .background(
+                                color = Color.LightGray.copy(alpha = 0.3f),
+                                shape = CircleShape
+                            )
+                    )
+                }
+            }
+        }
+
+        Button(
+            onClick = onGenerateClick,
+            modifier = Modifier.padding(top = 16.dp)
+        ) {
+            Text(stringResource(R.string.generate_new_number_set))
         }
     }
 }
